photo-local-server/requirements.txt
flask
flask-cors
google-auth
requests
python-dotenv